# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FreeWeight",
    "version": "1.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/evidence": {
      "get": {
        "description": "Return the current ``capability.evidence`` records (API §6).\n\nA collection envelope whose ``items`` are SetSpec envelopes — the two nest in exactly that\norder and never the reverse (ADR-0025 §2). ``user.*`` records carry ``goal_hash``,\n``score_method_mix``, ``judge_set``, ``calibration`` and ``judge_validity_factor``\n(ADR-0032 §5); a goal below its calibration gate has no record here at all.\n\nArgs:\n    request: The incoming request; the database handle lives on its application state.\n    capability: Exact capability ID.\n    model: Model canonical ID, ULID or unambiguous prefix.\n    machine: Machine fingerprint.\n    runtime_profile: Runtime profile hash.\n    min_confidence: Records at or above this confidence.\n    limit: Page size, clamped to 500.\n    cursor: Opaque continuation token.\n\nReturns:\n    The collection envelope: ``items`` and ``page``, and ``explanations`` — one per item, in\n    the same order: its staleness as of this request under this installation's policy, and\n    the confidence factors it was computed from. Beside the envelopes rather than inside them,\n    because staleness is a reading of a record at one instant and not part of the record.\n\nRaises:\n    ValidationError: The cursor was not issued here, or ``model`` is an ambiguous prefix.\n    NotFoundError: ``model`` matches nothing.",
        "operationId": "list_evidence_api_v1_evidence_get",
        "parameters": [
          {
            "description": "Exact capability ID, e.g. tool_use or user.house_voice.",
            "in": "query",
            "name": "capability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact capability ID, e.g. tool_use or user.house_voice.",
              "title": "Capability"
            }
          },
          {
            "description": "Model canonical ID, ULID or prefix.",
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Model canonical ID, ULID or prefix.",
              "title": "Model"
            }
          },
          {
            "description": "Machine fingerprint.",
            "in": "query",
            "name": "machine",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Machine fingerprint.",
              "title": "Machine"
            }
          },
          {
            "description": "Runtime profile hash.",
            "in": "query",
            "name": "runtime_profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Runtime profile hash.",
              "title": "Runtime Profile"
            }
          },
          {
            "description": "Records at or above this confidence.",
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Records at or above this confidence.",
              "title": "Min Confidence"
            }
          },
          {
            "description": "Page size; clamped to 500.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size; clamped to 500.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Evidence Api V1 Evidence Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List capability evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/v1/evidence/export": {
      "get": {
        "description": "Return one ``benchmark.evidence_bundle`` — the file LoadCoach imports (API §6, §10).\n\nA single SetSpec envelope with no collection wrapper, because a bundle is one document.\n``complete`` is ``true`` only when nothing narrows the selection: only a complete bundle lets\na consumer infer removals, and a filtered one must never be mistaken for it (ADR-0022 §5).\n\nArgs:\n    request: The incoming request.\n    since: RFC 3339; evidence whose ``computed_at`` is later, on FreeWeight's clock.\n    capability: Exact capability ID.\n    model: Model canonical ID, ULID or unambiguous prefix.\n    machine: Machine fingerprint.\n    runtime_profile: Runtime profile hash.\n    min_confidence: Records at or above this confidence.\n\nReturns:\n    The envelope as canonical JSON, with a ``Content-Disposition`` naming a file so a browser\n    saves it rather than rendering it.\n\nRaises:\n    ValidationError: ``since`` is malformed, or ``model`` is an ambiguous prefix.\n    NotFoundError: ``model`` matches nothing.",
        "operationId": "export_evidence_api_v1_evidence_export_get",
        "parameters": [
          {
            "description": "RFC 3339. Returns evidence whose computed_at is later, on FreeWeight's clock. Send back the generated_at of the bundle you received last time; never your own clock.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "RFC 3339. Returns evidence whose computed_at is later, on FreeWeight's clock. Send back the generated_at of the bundle you received last time; never your own clock.",
              "title": "Since"
            }
          },
          {
            "description": "Exact capability ID, e.g. tool_use or user.house_voice.",
            "in": "query",
            "name": "capability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact capability ID, e.g. tool_use or user.house_voice.",
              "title": "Capability"
            }
          },
          {
            "description": "Model canonical ID, ULID or prefix.",
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Model canonical ID, ULID or prefix.",
              "title": "Model"
            }
          },
          {
            "description": "Machine fingerprint.",
            "in": "query",
            "name": "machine",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Machine fingerprint.",
              "title": "Machine"
            }
          },
          {
            "description": "Runtime profile hash.",
            "in": "query",
            "name": "runtime_profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Runtime profile hash.",
              "title": "Runtime Profile"
            }
          },
          {
            "description": "Records at or above this confidence.",
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Records at or above this confidence.",
              "title": "Min Confidence"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the evidence bundle",
        "tags": [
          "evidence"
        ]
      }
    },
...
  }
}